Seems to fix the awful hack at least.

+++ term/TermInfoOutput.cpp
@@ -7,11 +7,11 @@ void TermInfoOutput::output(double video_pts, double audio_pts, bool show_video,
 {
    printf("\r");
    if (show_video)
-      printf("  V: %7.2lf", video_pts);
+      printf("  V: %7.2f", static_cast<float>(video_pts));
    if (show_audio)
-      printf("  A: %7.2lf", audio_pts);
+      printf("  A: %7.2f", static_cast<float>(audio_pts));
    if (show_video && show_audio)
-      printf("  Delta: %7.2lf", video_pts - audio_pts);
+      printf("  Delta: %7.2f", static_cast<float>(video_pts - audio_pts));
    printf("         ");
    fflush(stdout);
 }
